Optimization
Tasks:

Coding the Likelihood Function

Implement the likelihood function for an AR(7) model in Python. You are required to code both the conditional and unconditional likelihood functions.
Conditional Likelihood: This approach uses the first 7 observations as given and starts the likelihood calculation from the 8th observation.
Unconditional Likelihood: This approach integrates over the initial conditions (first 7 observations) using their theoretical or estimated distribution.
Use the following model specification for the AR(7) process:
where 
 is i.i.d. normal with mean zero and variance 
.

Maximizing the Likelihood

Write Python, Julia or R code to maximize the likelihood functions (both conditional and unconditional) with respect to the parameters 
 and 
. You may consider using optimization routines available in libraries such as scipy.optimize or Optim.jl.
Parameter Estimation

Estimate the parameters of the AR(7) model using both conditional and unconditional likelihood approaches. Utilize the monthly log differences of INDPRO from the FRED-MD dataset for this purpose.
You will need to preprocess the data to calculate log differences.
Forecasting

With the estimated parameters from both approaches, forecast the future values of the log differences of INDPRO for the next 8 months (1,2,......,8).
(Optional) Provide a brief comparison of the forecast accuracy from both sets of parameters based on out-of-sample forecasting. Discuss any notable differences and potential reasons for these differences.